Portfolio/Personal Projects/Discord Bots/Liguini/pyproject.toml
[tool.poetry]
name = "repl_python3_Liguini"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
replit = "^3.0.0"
ffmpeg = "^1.4"
PyNaCl = "^1.4.0"
xintool = "^0.0.7"
flask = "^1.1.2"
google-images-download = "^2.8.0"
selenium = "^3.141.0"
youtube_dl = "^2021.12.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
